add correct info in modules
