+++ js/functions.js
@@ -59,4 +59,4 @@ const getRandomInteger = (a, b) => {
   const result = Math.random() * (upper - lower + 1) + lower;
   return Math.floor(result);
 };
-getRandomInteger();
+export {getRandomInteger};
